requirements.txt
# Pokemon Multi-Agent System Dependencies
# Poke Agent - AI-Powered Pokemon TCG Dashboard

# Core server
flask>=3.0.0
flask-cors>=4.0.0
flask-compress>=1.14
requests>=2.31.0
gunicorn>=21.0.0

# Discord bot (multi-user support)
discord.py>=2.3.0
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Security (encrypted payment storage)
cryptography>=41.0.0

# Secure Payment Processing (cards never touch our servers)
stripe>=7.0.0                  # Stripe SDK for card payments

# Web scraping (for stock checking)
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Optional: AI Visual Grading
# openai>=1.0.0        # For GPT-4 Vision
# anthropic>=0.8.0     # For Claude Vision

# Optional: Multi-Channel Notifications
# twilio>=8.0.0        # For SMS notifications
# pushbullet.py>=0.12  # For Pushbullet push notifications
# python-telegram-bot>=20.0  # For Telegram notifications

# Browser Automation (for hybrid approach - fallback when requests blocked)
undetected-chromedriver>=3.5.0
selenium>=4.15.0

# Signal system (event-driven stock checking)
blinker>=1.7.0

# CAPTCHA Solving (optional - add API keys to .env)
# 2Captcha: https://2captcha.com (recommended)
# twocaptcha>=1.1.0
# Anti-Captcha: https://anti-captcha.com
# python3-anticaptcha>=1.0.0
